Move the WebClientConfig to configuration folder since it is used not only in "tests" Refactor application.properties into application.yml
