+++ role-service-client/src/main/java/uk/ac/ox/ndph/mts/roleserviceclient/configuration/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * Package documentation
+ */
+package uk.ac.ox.ndph.mts.roleserviceclient.configuration;
