--- role-service-client/src/main/resources/application.properties
@@ -1,10 +1,0 @@
-application.message="My app settings value here"
-role.service.url=<role-service-url>
-role.service.name=role-service
-role.service.endpoint.roles=/roles
-role.service.endpoint.exists=/roles/{id}
-role.service.endpoint.role=/roles/{id}
-role.service.endpoint.paged=/roles/
-role.service.endpoint.roles.by.ids=/roles
-role.service.endpoint.roles.create=/roles
-role.service.endpoint.update.permissions=/roles/{id}/permissions
