+++ role-service-client/src/main/resources/application.yml
@@ -1,0 +1,16 @@
+application:
+  message: '"My app settings value here"'
+role:
+  service:
+    name: role-service
+    url: <role-service-url>
+    endpoint:
+      exists: /roles/{id}
+      paged: /roles/
+      role: /roles/{id}
+      roles: /roles
+        create: /roles
+        by:
+          ids: roles
+      update:
+        permissions: /roles/{id}/permissions
