+++ role-service-client/src/test/java/uk/ac/ox/ndph/mts/roleserviceclient/GetPagedTest.java
@@ -9,7 +9,7 @@ import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.web.reactive.function.client.WebClient;
 import uk.ac.ox.ndph.mts.roleserviceclient.common.MockWebServerWrapper;
-import uk.ac.ox.ndph.mts.roleserviceclient.common.WebClientConfig;
+import uk.ac.ox.ndph.mts.roleserviceclient.configuration.WebClientConfig;
 
 import java.io.IOException;
