+++ role-service-client/src/test/java/uk/ac/ox/ndph/mts/roleserviceclient/GetRolesByIdsTest.java
@@ -8,7 +8,7 @@ import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.web.reactive.function.client.WebClient;
 import uk.ac.ox.ndph.mts.roleserviceclient.common.MockWebServerWrapper;
-import uk.ac.ox.ndph.mts.roleserviceclient.common.WebClientConfig;
+import uk.ac.ox.ndph.mts.roleserviceclient.configuration.WebClientConfig;
 import uk.ac.ox.ndph.mts.roleserviceclient.model.PermissionDTO;
 import uk.ac.ox.ndph.mts.roleserviceclient.model.RoleDTO;
 import uk.ac.ox.ndph.mts.roleserviceclient.exception.RestException;
