+++ role-service-client/src/test/java/uk/ac/ox/ndph/mts/roleserviceclient/RoleIdExistsTest.java
@@ -11,7 +11,7 @@ import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.http.HttpStatus;
 import org.springframework.web.reactive.function.client.WebClient;
 import uk.ac.ox.ndph.mts.roleserviceclient.common.MockWebServerWrapper;
-import uk.ac.ox.ndph.mts.roleserviceclient.common.WebClientConfig;
+import uk.ac.ox.ndph.mts.roleserviceclient.configuration.WebClientConfig;
 import uk.ac.ox.ndph.mts.roleserviceclient.exception.RestException;
 
 import java.io.IOException;
